refactor: validation 관련 함수 분리
